chore(convert_repo_json): move Cli and Emit to cli module
